+++ helper.js
@@ -83,7 +83,7 @@ function guidToBytes(guid) {
     var bytes = [];
     guid.split('-').map((number, index) => {
         var bytesInChar = index < 3 ? number.match(/.{1,2}/g).reverse() : number.match(/.{1,2}/g);
-        bytesInChar.map((byte) => { bytes.push(parseInt(byte, 16)); })
+        bytesInChar.map((byte) => { bytes.push(parseInt(byte, 16)); });
     });
     return bytes;
 }
@@ -178,6 +178,6 @@ helper.checkEnvVars = function () {
     }
 
     return env_check;
-}
+};
 
 module.exports = helper;